add max cluster size for foldmason condition to script

+++ data_creation_scripts/foldseek/create_foldseek_struct_from_db.py
@@ -39,7 +39,7 @@ def run_foldmason_on_pdbs(filelist, output_dir, tmp_dir):
         raise
 
 
-def save_pdbs_to_parquet(save_dir, pdbs_dir, clusters_to_save, parquet_id, metadata_lookup, run_foldmason=False):
+def save_pdbs_to_parquet(save_dir, pdbs_dir, clusters_to_save, parquet_id, metadata_lookup, run_foldmason=False, max_cluster_size_for_foldmason=None):
     # TODO: it would be cleaner for clusters_to_save values to be metadata-augmented dicts
     # Save the pdbs to parquet
     results = []
@@ -71,18 +71,23 @@ def save_pdbs_to_parquet(save_dir, pdbs_dir, clusters_to_save, parquet_id, metad
             
         # Run FoldMason on the cluster
         if run_foldmason:
-            foldmason_outdir = os.path.join(pdbs_dir, cluster_id)
-            os.makedirs(foldmason_outdir)
-            run_foldmason_on_pdbs(cluster_filelist, foldmason_outdir, foldmason_outdir)
+            if max_cluster_size_for_foldmason is not None and len(cluster_filelist) > max_cluster_size_for_foldmason:
+                print(f"Skipping FoldMason for {cluster_id} due to size {len(cluster_filelist)}")
+                has_foldmason_results = False
+            else:
+                foldmason_outdir = os.path.join(pdbs_dir, cluster_id)
+                os.makedirs(foldmason_outdir)
+                run_foldmason_on_pdbs(cluster_filelist, foldmason_outdir, foldmason_outdir)
 
-            # Read AA and 3Di alignments, skip the accessions
-            labels, msta_seqs = read_fasta(os.path.join(foldmason_outdir, "result_aa.fa"))
-            perm = [labels.index(afdb_id) for afdb_id in cluster_members]
-            msta_seqs = [msta_seqs[ix] for ix in perm]
-            struct_labels, msta_3di = read_fasta(os.path.join(foldmason_outdir, "result_3di.fa"))
-            assert labels == struct_labels
-            msta_3di = [msta_3di[ix] for ix in perm]
-            shutil.rmtree(foldmason_outdir)
+                # Read AA and 3Di alignments, skip the accessions
+                labels, msta_seqs = read_fasta(os.path.join(foldmason_outdir, "result_aa.fa"))
+                perm = [labels.index(afdb_id) for afdb_id in cluster_members]
+                msta_seqs = [msta_seqs[ix] for ix in perm]
+                struct_labels, msta_3di = read_fasta(os.path.join(foldmason_outdir, "result_3di.fa"))
+                assert labels == struct_labels
+                msta_3di = [msta_3di[ix] for ix in perm]
+                shutil.rmtree(foldmason_outdir)
+                has_foldmason_results = True
 
         for pdb in cluster_filelist:
             os.remove(pdb)
@@ -99,7 +104,7 @@ def save_pdbs_to_parquet(save_dir, pdbs_dir, clusters_to_save, parquet_id, metad
             "accessions": accessions,
             "af50_cluster_id": af50_cluster_id,
         }
-        if run_foldmason:
+        if has_foldmason_results:
             res["msta_seqs"] = msta_seqs
             res["msta_3di"] = msta_3di
         results.append(res)
@@ -178,6 +183,7 @@ def create_foldseek_parquets(
     af50_representative_only=False,
     show_tqdm=False,
     run_foldmason=False,
+    max_cluster_size_for_foldmason=None,
 ):
     # TODO: instead of loading the cluster dictionary we can just save a file which lists the cluster sizes.
     # af50 version doesn't really work with parquet ids...no i guess it still does: db is limited to a single parquet in that case. 
@@ -242,6 +248,7 @@ def create_foldseek_parquets(
             clusters_to_save=parquet_cluster_membership,
             parquet_id=parquet_id,
             metadata_lookup=parquet_metadata_lookup,
+            max_cluster_size_for_foldmason=max_cluster_size_for_foldmason,
             run_foldmason=run_foldmason and not (representative_only or af50_representative_only),
         )
     shutil.rmtree(os.path.join(scratch_dir, job_prefix))
@@ -259,6 +266,7 @@ if __name__ == "__main__":
     parser.add_argument("--save_dir", default=None)
     parser.add_argument("--representative_only", action="store_true")
     parser.add_argument("--af50_representative_only", action="store_true")
+    parser.add_argument("--max_cluster_size_for_foldmason", type=int, default=None)
     args = parser.parse_args()
 
     if args.save_dir is None:
@@ -287,4 +295,5 @@ if __name__ == "__main__":
         run_foldmason=args.run_foldmason,
         representative_only=args.representative_only,
         af50_representative_only=args.af50_representative_only,
+        max_cluster_size_for_foldmason=args.max_cluster_size_for_foldmason,
     )
